+++ src/core/tamper/uninitializedvariable.py
@@ -19,7 +19,7 @@ import string
 from src.utils import settings
 
 """
-About: Adds (randomly generated) uninitialized bash variables, between the characters of each command in a given payload.
+About: Adds (randomly generated) uninitialized bash variables between the characters of each command in a given payload.
 Notes: This tamper script works against Unix-like target(s).
 Reference: https://www.secjuice.com/web-application-firewall-waf-evasion/
 """